opt: i18n config
